+++ src/main/java/duke/Storage.java
@@ -1,6 +1,7 @@
 package duke;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -8,6 +9,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 
 import duke.tasks.Deadline;
 import duke.tasks.Event;
@@ -44,8 +46,94 @@ public class Storage {
      *
      * @return the save file.
      */
-    public File load() {
-        return this.data;
+    public ArrayList<Task> load() throws FileNotFoundException {
+        ArrayList<Task> tasks = new ArrayList<>();
+
+        Scanner dataScanner = new Scanner(data);
+        while (dataScanner.hasNext()) {
+            String dataLine = dataScanner.nextLine();
+            String[] split = dataLine.split(" ");
+            String type = split[0];
+            boolean taskDone = split[2].equals("1"); //if taskDone = 1, task was done
+            String desc;
+
+            switch (type) {
+            case "T":
+                StringBuilder todoBuilder = new StringBuilder();
+                for (int i = 4; i < split.length; i++) {
+                    if (i != 4) {
+                        todoBuilder.append(" ");
+                    }
+                    todoBuilder.append(split[i]);
+                }
+                desc = todoBuilder.toString();
+
+                tasks.add(new Todo(desc, taskDone));
+                break;
+
+            case "D":
+                StringBuilder deadlineBuilder = new StringBuilder();
+                StringBuilder byBuilder = new StringBuilder();
+                String by;
+                boolean byFound = false;
+
+                for (int i = 4; i < split.length; i++) {
+                    if (byFound) {
+                        if (!byBuilder.toString().equals("")) {
+                            byBuilder.append(" ");
+                        }
+                        byBuilder.append(split[i]);
+                    } else {
+                        if (i == 4) {
+                            deadlineBuilder.append(split[i]);
+                        } else if (split[i].equals("|")) {
+                            byFound = true;
+                        } else {
+                            deadlineBuilder.append(" ");
+                            deadlineBuilder.append(split[i]);
+                        }
+                    }
+                }
+                desc = deadlineBuilder.toString();
+                by = byBuilder.toString();
+
+                tasks.add(new Deadline(desc, by, taskDone));
+                break;
+
+            case "E":
+                StringBuilder eventBuilder = new StringBuilder();
+                StringBuilder atBuilder = new StringBuilder();
+                String at;
+                boolean atFound = false;
+
+                for (int i = 4; i < split.length; i++) {
+                    if (atFound) {
+                        if (!atBuilder.toString().equals("")) {
+                            atBuilder.append(" ");
+                        }
+                        atBuilder.append(split[i]);
+                    } else {
+                        if (i == 4) {
+                            eventBuilder.append(split[i]);
+                        } else if (split[i].equals("|")) {
+                            atFound = true;
+                        } else {
+                            eventBuilder.append(" ");
+                            eventBuilder.append(split[i]);
+                        }
+                    }
+                }
+                desc = eventBuilder.toString();
+                at = atBuilder.toString();
+
+                tasks.add(new Event(desc, at, taskDone));
+                break;
+
+            default:
+
+            }
+        }
+        return tasks;
     }
 
     /**
@@ -54,7 +142,7 @@ public class Storage {
      * @param task the task to be added to the save file.
      * @throws IOException when an IO operation fails.
      */
-    public void add(Task task) throws IOException {
+    public void save(Task task) throws IOException {
         FileWriter fw = new FileWriter("data/tasks.txt", true);
         if (task instanceof Todo) {
             Todo todo = (Todo) task;
