+++ src/main/java/duke/commands/AddTodoCommand.java
@@ -34,27 +34,22 @@ public class AddTodoCommand extends Command {
     /**
      * Executes the command. Adds deadline to task list. Updates the save file.
      *
-     * @param tasks   the task list.
-     * @param storage the storage of the programme.
+     * @param tasks the task list.
      */
     @Override
-    public void execute(TaskList tasks, Storage storage) {
-        try {
-            FileWriter fw = new FileWriter("data/tasks.txt", true);
-            Todo todo = new Todo(super.getDesc(), false);
-            tasks.add(todo);
-            storage.add(todo);
+    public void execute(TaskList tasks) throws IOException {
 
-            System.out.println("Got it. I've added this task:");
-            System.out.println(todo);
+        Todo todo = new Todo(super.getDesc(), false);
 
-            if (tasks.size() == 1) {
-                System.out.println("Now you have 1 task in the list.");
-            } else {
-                System.out.println("Now you have " + tasks.size() + " tasks in the list.");
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
+        tasks.add(todo);
+
+        System.out.println("Got it. I've added this task:");
+        System.out.println(todo);
+
+        if (tasks.size() == 1) {
+            System.out.println("Now you have 1 task in the list.");
+        } else {
+            System.out.println("Now you have " + tasks.size() + " tasks in the list.");
         }
     }
 }
