+++ src/main/java/duke/commands/DeleteCommand.java
@@ -35,28 +35,20 @@ public class DeleteCommand extends Command {
 
     /**
      * Executes the command. Adds deadline to task list. Updates the save file.
-     *
-     * @param tasks   the task list.
-     * @param storage the storage of the programme.
      */
     @Override
-    public void execute(TaskList tasks, Storage storage) {
+    public void execute(TaskList tasks) throws IOException {
         System.out.println("-------------------------------------");
         System.out.println("Very well, Master Wayne. This task has been deleted as per your request.");
         System.out.println((deleteIndex) + ". " + tasks.get(deleteIndex - 1)); //actual index is index - 1
 
         tasks.remove(deleteIndex - 1);
+
         if (tasks.size() == 1) {
             System.out.println("Now you have 1 task in the list.");
         } else {
             System.out.println("Now you have " + tasks.size() + " tasks in the list.");
         }
         System.out.println("-------------------------------------");
-
-        try {
-            storage.delete(deleteIndex);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
 }
