+++ text-ui-test/data/tasks.txt
@@ -1,1 +1,3 @@
-
+E | 0 | project meeting | Aug 6th 2-4pm
+T | 0 | join sports club
+T | 0 | read book
