+++ text-ui-test/runtest.bat
@@ -5,11 +5,10 @@ if not exist ..\bin mkdir ..\bin
 
 REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
+break>data\tasks.txt
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java\duke -Xlint:none -d ..\bin ..\src\main\java\duke\*.java
-javac  -cp ..\src\main\java\duke\commands -Xlint:none -d ..\bin ..\src\main\java\duke\commands\*.java
-javac  -cp ..\src\main\java\duke\tasks -Xlint:none -d ..\bin ..\src\main\java\duke\tasks\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\duke\*.java
 
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
@@ -18,7 +17,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Duke < input.txt > ACTUAL.TXT
+java -classpath ..\bin duke.Duke < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
